- Practical test
A very popular online gaming community has 100000+ users playing at a given time. We need to design a scoreboard which lists the top ten scores along with player names. Implement a basic iPhone UI to display this list. Please design for optimal performance so that the scoreboard is as real-time as possible. As millions of users engage in a gaming tournament simultaneously, scoreboard needs to be updated every time any score gets updated. If not implemented properly the updates list will be so heavy and so frequent that the program might get too slow and might just fail the entire experience